SSHClient: set Transport _preferred_keys if have server hostkey
If we have server host keys that will be checked, need need to
negotiate for one of the types we have. Commonly, openssh could
have saved an ecdsa key but SSHClient will let the Transport
negotiate for an rsa key.

Then it would consider a key of a non-corresponding type to be "missing".
That situation is also now a BadHostKeyException.